APIGOV-22856 - update to codegen for quotas (#483)
* APIGOV-22856 - code gen, update float32 to float64

* APIGOV-22856 - update to allow nil quota on access request

* APIGOV-22856 - update for nil quota on access requests

+++ pkg/apic/provisioning/quota_test.go
@@ -37,7 +37,7 @@ func TestNewQuotaFromAccessRequest(t *testing.T) {
 			ar := v1alpha1.NewAccessRequest("name", "environment")
 			if tt.intervalString != "" {
 				ar.Spec.Quota = &v1alpha1.AccessRequestSpecQuota{
-					Limit:    float32(tt.limit),
+					Limit:    float64(tt.limit),
 					Interval: tt.intervalString,
 				}
 			}
